Moves the addGroovyProperties and removeGroovyProperties to the class WpsServerImpl.
