+++ bundles/wps/wps-service-orbisgis/src/main/java/org/orbisgis/wpsserviceorbisgis/OrbisGISWpsServer.java
@@ -105,20 +105,4 @@ public interface OrbisGISWpsServer extends WpsServer {
      * @return The list of the available SRID.
      */
     List<String> getSRIDList();
-
-    /**
-     * Adds to the server execution properties which will be set to the GroovyObject for the execution.
-     * Those properties will be accessible inside the groovy script as variables which name is the map entry key.
-     * For example :
-     * If the propertiesMap contains <"message", "HelloWorld">, inside the groovy script you can print the message this
-     * way : 'print message'
-     * @param propertiesMap Map containing the properties to be passed to the GroovyObject
-     */
-    void addGroovyProperties(Map<String, Object> propertiesMap);
-
-    /**
-     * Removes the properties already set for the GroovyObject for the execution.
-     * @param propertiesMap Map containing the properties to be removed
-     */
-    void removeGroovyProperties(Map<String, Object> propertiesMap);
 }
